Set up post and get for Admin and Creator databases. Ready for front end calls

+++ backEnd/src/main/java/com/wellnessapp/demo/AdminRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface AdminRepository extends MongoRepository<Admin, Long> {
+public interface AdminRepository extends MongoRepository<Admin, Integer> {
     // possible functions for querrying the db
     public Admin findByUsername(String username);
     public List<Admin> findByPhone(String phone);
